Initial commit.. Full working application for taking preferences
A full working slack app for taking the preferences of members
in the Channel. All choices will get saved and can be accessed by
sending "total" in paramter of command.
Message will be automatically deleted after 15 minutes of posting in Channel.

To create a new form for taking preferences, write the app access token in .token.txt file
and password in .password.txt file of private directory.
User will need password for requesting a new form for taking choices of member.